# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AI Card Review Web App API
  version: 0.1.0
  description: >
    Contract for the separated frontend/backend AI card review application. The API exposes
    learner workspace operations for material intake, knowledge curation, AI draft workflows,
    deck/card management, review sessions, review plans, statistics, and exports/imports.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /review/sessions:
    post:
      tags: [Review]
      summary: Start a direct or plan-based review session
      operationId: startReviewSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartReviewSessionRequest'
      responses:
        '201':
          description: Started review session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewSession'
  /review/sessions/{sessionId}/answers:
    post:
      tags: [Review]
      summary: Record one card answer in a review session
      operationId: recordReviewAnswer
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordReviewAnswerRequest'
      responses:
        '201':
          description: Recorded answer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResult'
  /review/sessions/{sessionId}:
    patch:
      tags: [Review]
      summary: Pause, resume, complete, or abandon a review session
      operationId: updateReviewSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReviewSessionRequest'
      responses:
        '200':
          description: Updated review session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewSession'
  /plans:
    get:
      tags: [Plans]
      summary: List review plans
      operationId: listPlans
      parameters:
        - $ref: '#/components/parameters/StatusFilter'
      responses:
        '200':
          description: Review plans
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReviewPlan'
    post:
      tags: [Plans]
      summary: Create a manual review plan
      operationId: createPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlanRequest'
      responses:
        '201':
          description: Created review plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewPlan'
  /plans/generate:
    post:
      tags: [Plans, AI Jobs]
      summary: Generate an AI-assisted review plan draft
      operationId: generatePlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneratePlanRequest'
      responses:
        '202':
          description: Plan generation job queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAccepted'
  /plans/{planId}:
    patch:
      tags: [Plans]
      summary: Update plan metadata, schedule, or status
      operationId: updatePlan
      parameters:
        - $ref: '#/components/parameters/PlanId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlanRequest'
      responses:
        '200':
          description: Updated review plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewPlan'
  /plans/{planId}/optimize:
    post:
      tags: [Plans, AI Jobs]
      summary: Queue AI-assisted or rules-based plan optimization
      operationId: optimizePlan
      parameters:
        - $ref: '#/components/parameters/PlanId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptimizePlanRequest'
      responses:
        '202':
          description: Plan optimization job queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAccepted'
  /plans/{planId}/revisions:
    get:
      tags: [Plans]
      summary: List plan revisions
      operationId: listPlanRevisions
      parameters:
        - $ref: '#/components/parameters/PlanId'
      responses:
        '200':
          description: Plan revisions
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PlanRevision'
  /plans/{planId}/revisions/{revisionId}/restore:
    post:
      tags: [Plans]
      summary: Restore a compatible prior plan revision
      operationId: restorePlanRevision
      parameters:
        - $ref: '#/components/parameters/PlanId'
        - name: revisionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Restored review plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewPlan'
  /statistics:
    get:
      tags: [Statistics]
      summary: Get review statistics by scope and period
      operationId: getStatistics
      parameters:
        - name: scopeType
          in: query
          schema:
            type: string
            enum: [workspace, tag, deck, plan, card]
        - name: scopeId
          in: query
          schema:
            type: string
        - name: periodStart
          in: query
          schema:
            type: string
            format: date
        - name: periodEnd
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Statistics snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewStatistics'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
  parameters:
...
    StatusFilter:
      name: status
      in: query
      schema:
        type: string
...
    PlanId:
      name: planId
      in: path
      required: true
      schema:
        type: string
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    JobAccepted:
      type: object
      required: [job]
      properties:
        job:
          $ref: '#/components/schemas/AIJob'
...
    AIJob:
      type: object
      required: [id, jobType, status, progressPercent, currentStep]
      properties:
        id:
          type: string
        jobType:
          type: string
          enum: [material_analysis, card_generation, plan_generation, plan_optimization, import, export]
        status:
          type: string
          enum: [queued, running, partial_success, succeeded, failed, cancelled]
        progressPercent:
          type: integer
          minimum: 0
          maximum: 100
        currentStep:
          type: string
        errorMessage:
          type: string
        resultDraftIds:
          type: array
          items:
            type: string
...
    StartReviewSessionRequest:
      type: object
      required: [sessionType]
      properties:
        sessionType:
          type: string
          enum: [direct, plan_based]
        deckId:
          type: string
        planId:
          type: string
        dueOnly:
          type: boolean
          default: false
        maxCards:
          type: integer
    ReviewSession:
      type: object
      required: [id, sessionType, status, cardsRemaining]
      properties:
        id:
          type: string
        sessionType:
          type: string
          enum: [direct, plan_based]
        status:
          type: string
          enum: [active, paused, completed, abandoned]
        cardsRemaining:
          type: integer
        summary:
          type: string
    RecordReviewAnswerRequest:
      type: object
      required: [cardId, result, elapsedResponseMs]
      properties:
        cardId:
          type: string
        result:
          type: string
          enum: [again, hard, good, easy, correct, incorrect]
        confidence:
          type: integer
          minimum: 1
          maximum: 5
        elapsedResponseMs:
          type: integer
    ReviewResult:
      type: object
      required: [id, cardId, result, nextDueAt]
      properties:
        id:
          type: string
        cardId:
          type: string
        result:
          type: string
        nextDueAt:
          type: string
          format: date-time
        scheduleReason:
          type: string
    UpdateReviewSessionRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [pause, resume, complete, abandon]
    CreatePlanRequest:
      type: object
      required: [name, deckIds, startDate, targetCompletionDate, dailyCapacity]
      properties:
        name:
          type: string
        goal:
          type: string
        deckIds:
          type: array
          items:
            type: string
        startDate:
          type: string
          format: date
        targetCompletionDate:
          type: string
          format: date
        dailyCapacity:
          type: integer
        overlapPolicy:
          type: string
          enum: [prevent, allow_with_warning]
          default: prevent
    GeneratePlanRequest:
      allOf:
        - $ref: '#/components/schemas/CreatePlanRequest'
        - type: object
          required: [promptText]
          properties:
            promptPresetId:
              type: string
            promptText:
              type: string
    UpdatePlanRequest:
      type: object
      properties:
        name:
          type: string
        goal:
          type: string
        status:
          type: string
          enum: [active, paused, completed, archived]
        dailyCapacity:
          type: integer
        targetCompletionDate:
          type: string
          format: date
    OptimizePlanRequest:
      type: object
      required: [mode]
      properties:
        mode:
          type: string
          enum: [ai_assisted, rules_based]
        promptPresetId:
          type: string
        promptText:
          type: string
        respectDailyCapacity:
          type: boolean
          default: true
    ReviewPlan:
      type: object
      required: [id, name, status, dailyCapacity]
      properties:
        id:
          type: string
        name:
          type: string
        goal:
          type: string
        status:
          type: string
          enum: [draft, active, paused, completed, archived]
        startDate:
          type: string
          format: date
        targetCompletionDate:
          type: string
          format: date
        dailyCapacity:
          type: integer
        deckIds:
          type: array
          items:
            type: string
        currentRevisionId:
          type: string
    PlanRevision:
      type: object
      required: [id, revisionNumber, changeSource, changeSummary, createdAt]
      properties:
        id:
          type: string
        revisionNumber:
          type: integer
        changeSource:
          type: string
          enum: [manual, ai_assisted, rules_based]
        changeSummary:
          type: string
        changeReason:
          type: string
        createdAt:
          type: string
          format: date-time
    ReviewStatistics:
      type: object
      required: [scopeType, cardsReviewed, recallRate, overdueCount, weakAreas]
      properties:
        scopeType:
          type: string
        scopeId:
          type: string
        cardsReviewed:
          type: integer
        recallRate:
          type: number
        overdueCount:
          type: integer
        completionRate:
          type: number
        weakAreas:
          type: array
          items:
            type: object
            required: [label, score]
            properties:
              label:
                type: string
              score:
                type: number
...